Package is behaving weirdly with setup.py develop.
The libraries -- in fact the whole setup.cfg -- does not seem to be used
when running setup.py develop (e.g., pip install -e). I noticed this
when I saw undefined sqlite3 symbols after trying to import the C
extension. A fix seems to be to just explicitly specify the libraries
when building against the system sqlite.