Increase the host parameter limit
This increases the maximum number of placeholders SQLite will accept.
See: https://www.sqlite.org/limits.html#max_variable_number

By default it is set at 999, presumably because the default
configuration needs to work with highly memory-constrained embedded
devices. Most OS distributions of SQLite increase this number
significantly. For example, on Ubuntu it is 250,000 and on OS X it is
500,000. See:
https://launchpad.net/ubuntu/+source/sqlite3/3.8.0.2-1ubuntu1
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=717900