Increase maximum mmap size
This increases the maximum size of database file which can be
memory-mapped from 2GB to 1TB. Note that this only changes the *maximum*
and the *default* remains at 0 (i.e. memory-mapping disabled) so this
shouldn't make any difference to users unless they explicitly ask to
mmap more than 2GB of data.

I'm not sure why the default is so low to be honest. Here are a few
threads discussing the benefits of increasing it:
http://sqlite.1065341.n5.nabble.com/SQLITE-MAX-MMAP-SIZE-2GB-default-td106789.html
https://github.com/xerial/sqlite-jdbc/issues/327
https://github.com/xerial/sqlite-jdbc/pull/328